requirements.txt
aiogram>=3.0,<4.0
openai>=2.0.0
python-dotenv>=1.0.0,<2.0.0
